+++ app/Http/Controllers/Admin/RoomTypeController.php
@@ -16,7 +16,7 @@ class RoomTypeController extends Controller
     public function index()
     {
         $data = RoomType::all();
-        return view('admin.roomType.index',['data'=>$data]);
+        return view('admin.roomType.index', ['data' => $data]);
     }
 
     /**
@@ -37,26 +37,25 @@ class RoomTypeController extends Controller
         $request->validate([
             'title' => 'required',
             'details' => 'required',
-            'price' => 'required', 
+            'price' => 'required',
         ]);
         $data = new RoomType;
         $data->title = $request->title;
         $data->price = $request->price;
         $data->details = $request->details;
-        if($request->imgs){
-        foreach($request->file('imgs') as $img){
-            $imgPath = $img->store('RoomTypeImages','public');
-            $imgData = new RoomTypeImage;
-            $imgData->room_type_id=$data->id;
-            $imgData->img_src=$imgPath;
-            $imgData->img_alt=$request->title;
-            $imgData->save();
-        }
+        if ($request->imgs) {
+            foreach ($request->file('imgs') as $img) {
+                $imgPath = $img->store('RoomTypeImages', 'public');
+                $imgData = new RoomTypeImage;
+                $imgData->room_type_id = $data->id;
+                $imgData->img_src = $imgPath;
+                $imgData->img_alt = $request->title;
+                $imgData->save();
+            }
         }
 
         $data->save();
-        return redirect('admin/roomtype')->with('success','Room Type has been Created Successfully!');
-        
+        return redirect('admin/roomtype')->with('success', 'Room Type has been Created Successfully!');
     }
 
     /**
@@ -66,7 +65,10 @@ class RoomTypeController extends Controller
     {
         //
         $data = RoomType::find($id);
-        return view('admin.roomType.show',['data'=>$data]);
+        if ($data == null) {
+            return redirect()->route('admin.roomtype.index')->with('danger', 'Not Found!');
+        }
+        return view('admin.roomType.show', ['data' => $data]);
     }
 
     /**
@@ -76,7 +78,10 @@ class RoomTypeController extends Controller
     {
         //
         $data = RoomType::find($id);
-        return view('admin.roomType.edit',['data'=>$data]);
+        if ($data == null) {
+            return redirect()->route('admin.roomtype.index')->with('danger', 'Not Found!');
+        }
+        return view('admin.roomType.edit', ['data' => $data]);
     }
 
     /**
@@ -92,18 +97,17 @@ class RoomTypeController extends Controller
             'price' => 'required',
         ]);
         $data->update($formFields);
-        if($request->hasFile('imgs')){
-            foreach($request->file('imgs') as $img){
-                $imgPath = $img->store('RoomTypeImages','public');
+        if ($request->hasFile('imgs')) {
+            foreach ($request->file('imgs') as $img) {
+                $imgPath = $img->store('RoomTypeImages', 'public');
                 $imgData = new RoomTypeImage;
-                $imgData->room_type_id=$data->id;
-                $imgData->img_src=$imgPath;
-                $imgData->img_alt=$request->title;
+                $imgData->room_type_id = $data->id;
+                $imgData->img_src = $imgPath;
+                $imgData->img_alt = $request->title;
                 $imgData->save();
             }
         }
-        return redirect('admin/roomtype')->with('success','Data has been updated Successfully!');
-        
+        return redirect('admin/roomtype')->with('success', 'Data has been updated Successfully!');
     }
 
     /**
@@ -112,30 +116,32 @@ class RoomTypeController extends Controller
     public function destroy($id)
     {
         $data = RoomType::find($id);
+        if ($data == null) {
+            return redirect()->route('admin.roomtype.index')->with('danger', 'Not Found!');
+        }
         //Delete It's Images
-            foreach ($data->roomtypeimages as $img){
-                $dataDelete = RoomTypeImage::find($img->id);
-                $path = 'storage/'.$dataDelete->img_src;
-                if(File::exists($path)) {
-                    File::delete($path);
-                }
-                //Delete Image
-                $dataDelete->delete();
+        foreach ($data->roomtypeimages as $img) {
+            $dataDelete = RoomTypeImage::find($img->id);
+            $path = 'storage/' . $dataDelete->img_src;
+            if (File::exists($path)) {
+                File::delete($path);
             }
+            //Delete Image
+            $dataDelete->delete();
+        }
         $data->delete();
-        return redirect('admin/roomtype')->with('danger','Data has been deleted Successfully!');
-
+        return redirect('admin/roomtype')->with('danger', 'Data has been deleted Successfully!');
     }
     public function destroy_image($id)
     {
         $data = RoomTypeImage::find($id);
         //Delete Image from storage
-        $path = 'storage/'.$data->img_src;
-        if(File::exists($path)) {
+        $path = 'storage/' . $data->img_src;
+        if (File::exists($path)) {
             File::delete($path);
         }
         //Delete Image
         $data->delete();
-        return response()->json(['bool'=>true]);
+        return response()->json(['bool' => true]);
     }
 }
